requirement.txt
asgiref==3.4.1
Django==3.2.8
django-redis==5.0.0
django-redis-cache==3.0.0
django-redis-sessions==0.6.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
